+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 - 0.2.9
   - (TODO) change sign in gradient simulation - flips orientation of old reconstructions
   - Included util file in mr0
-  - (TODO) Improved plot function to support pypulseq 1.4.0-dev
+  - Improved plot function to support pypulseq 1.4.0-dev
+  - (TODO) Add util to documentation
 - 0.2.8
   - Updated playground notebook to use new FOV definitions
   - Fixed SimData.recover()